First pass at implementation of authentication.
Most code rescued from other webtech apps, rpitv development, or the
in-development eclub site.

Uses a mixin to ActionController::Base to add storage and retrieval of
user info from session. Also adds the hooks for authenticates_access.

SessionController implemented, which deals with logging in and logging
out users, relying on rubycas-client for authentication, and looking up
data from a Person model.

The above changes are a rough-in implementation, expect rapid and
drastic changes, especially to the way the user model is handled.