Accept generic User model.
When the mixin is called in application controller, the programmer will
specify a model for authentication that the accessor will be an
instance of.  The user will all specify the field that corresponds to
the username returned from CAS, and optionally the field or method that
is to be used to determine whether a user is an admin.

We identify the accessor by the username, which we store in
session[:username].