Get rid of debug stuff in session/show.
It is now a place that it would be ok for a user to see: it indicates
the users login status (and name), and provides login/logout links. Also
a potential for admins to get session information if they need to debug
something quick.

It may be blanked out by placing an empty view in the host application's
app/views/sesssion/show.html.erb