Added sudo feature when enabled
Administrators can sudo (if there are administrators, and sudo is
enabled in the authenticate_rpi call), that is, pretend to be a
different user, by visiting session/sudo. If pretending to be a
non-admin, they get only the rights that that user has, except that they
can still sudo to be themselves or another user.